requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6 